Service with work every 3 minutes
