+++ Sunshine/app/src/main/java/com/cdesign/sunshine/sync/SunshineSyncService.java
@@ -1,0 +1,33 @@
+package com.cdesign.sunshine.sync;
+
+import android.app.Service;
+import android.content.Intent;
+import android.os.IBinder;
+import android.support.annotation.Nullable;
+import android.util.Log;
+
+import com.cdesign.sunshine.sync.SunshineSyncAdapter;
+
+/**
+ * Created by RealXaker on 18.08.2016.
+ */
+public class SunshineSyncService extends Service {
+    private static final Object sSyncAdapterLock = new Object();
+    private static SunshineSyncAdapter sSunshineSyncAdapter = null;
+
+    @Override
+    public void onCreate() {
+        Log.d("SunshineSyncService", "onCreate - SunshineSyncService");
+        synchronized (sSyncAdapterLock) {
+            if (sSunshineSyncAdapter == null) {
+                sSunshineSyncAdapter = new SunshineSyncAdapter(getApplicationContext(), true);
+            }
+        }
+    }
+
+    @Nullable
+    @Override
+    public IBinder onBind(Intent intent) {
+        return sSunshineSyncAdapter.getSyncAdapterBinder();
+    }
+}
