+++ Sunshine/app/src/main/java/com/cdesign/sunshine/ui/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.example.android.sunshine.app;
+package com.cdesign.sunshine.ui;
 
 import android.annotation.TargetApi;
 import android.content.Intent;
@@ -8,6 +8,8 @@ import android.preference.ListPreference;
 import android.preference.Preference;
 import android.preference.PreferenceActivity;
 import android.preference.PreferenceManager;
+
+import com.example.android.sunshine.app.R;
 
 public class SettingsActivity extends PreferenceActivity implements Preference.OnPreferenceChangeListener {
