--- Sunshine/app/src/main/java/com/example/android/sunshine/app/data/FetchWeatherTask.java
@@ -1,163 +1,0 @@
-package com.example.android.sunshine.app.data;
-
-import android.content.ContentUris;
-import android.content.ContentValues;
-import android.content.Context;
-import android.database.Cursor;
-import android.net.Uri;
-import android.os.AsyncTask;
-import android.util.Log;
-
-import com.example.android.sunshine.app.BuildConfig;
-import com.example.android.sunshine.app.data.db.WeatherContract;
-import com.example.android.sunshine.app.utils.ConstantManager;
-import com.example.android.sunshine.app.utils.Utils;
-
-import org.json.JSONException;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-/**
- * Created by Ageev Evgeny on 09.08.2016.
- */
-public class FetchWeatherTask extends AsyncTask<String, Void, Void> {
-    private final String LOG_TAG = FetchWeatherTask.class.getSimpleName();
-    //private ForecastAdapter mForecastAdapter;
-    private final Context mContext;
-
-    private boolean DEBUG = true;
-
-    public FetchWeatherTask(Context context) {
-        mContext = context;
-    }
-
-    private long addLocation(String locationSetting, String cityName, double lat, double lon) {
-        long locationId;
-
-        // First, check if the location with this city name exists in the db
-        Cursor locationCursor = mContext.getContentResolver().query(
-                WeatherContract.LocationEntry.CONTENT_URI,
-                new String[]{WeatherContract.LocationEntry._ID},
-                WeatherContract.LocationEntry.COLUMN_LOCATION_SETTING + " = ?",
-                new String[]{locationSetting},
-                null);
-
-        Log.d(LOG_TAG, "We Are Here!");
-        if (locationCursor.moveToFirst()) {
-            int locationIdIndex = locationCursor.getColumnIndex(WeatherContract.LocationEntry._ID);
-            locationId = locationCursor.getLong(locationIdIndex);
-            Log.d(LOG_TAG, "location from bd - " + locationIdIndex + ", " + locationId);
-        } else {
-            // Now that the content provider is set up, inserting rows of data is pretty simple.
-            // First create a ContentValues object to hold the data you want to insert.
-            ContentValues locationValues = new ContentValues();
-
-            // Then add the data, along with the corresponding name of the data type,
-            // so the content provider knows what kind of value is being inserted.
-            locationValues.put(WeatherContract.LocationEntry.COLUMN_CITY_NAME, cityName);
-            locationValues.put(WeatherContract.LocationEntry.COLUMN_LOCATION_SETTING, locationSetting);
-            locationValues.put(WeatherContract.LocationEntry.COLUMN_COORD_LAT, lat);
-            locationValues.put(WeatherContract.LocationEntry.COLUMN_COORD_LONG, lon);
-
-            // Finally, insert location data into the database.
-            Uri insertedUri = mContext.getContentResolver().insert(
-                    WeatherContract.LocationEntry.CONTENT_URI,
-                    locationValues
-            );
-
-            // The resulting URI contains the ID for the row.  Extract the locationId from the Uri.
-            locationId = ContentUris.parseId(insertedUri);
-            Log.d(LOG_TAG, "location writes to bd - " + locationId);
-        }
-
-        locationCursor.close();
-        // Wait, that worked?  Yes!
-        return locationId;
-    }
-
-    @Override
-    protected Void doInBackground(String... params) {
-        // If there's no zip code, there's nothing to look up.  Verify size of params.
-        if (params.length == 0) return null;
-        String locationQuery = params[0];
-
-        HttpURLConnection urlConnection = null;
-        BufferedReader reader = null;
-
-        String forecastJsonStr = null;
-        String format = "json";
-        String units = "metric";
-        int numDays = 14;
-
-        try {
-            //http://openweathermap.org/API#forecast
-            Uri builtUri = Uri.parse(ConstantManager.FORECAST_BASE_URI).buildUpon()
-                    .appendQueryParameter(ConstantManager.QUERY_PARAM, locationQuery)
-                    .appendQueryParameter(ConstantManager.FORMAT_PARAM, format)
-                    .appendQueryParameter(ConstantManager.UNITS_PARAM, units)
-                    .appendQueryParameter(ConstantManager.DAYS_PARAM, Integer.toString(numDays))
-                    .appendQueryParameter(ConstantManager.APPID_PARAM, BuildConfig.OPEN_WEATHER_MAP_API_KEY)
-                    .build();
-
-            URL url = new URL(builtUri.toString());
-            Log.d(LOG_TAG, "Built URI - " + builtUri.toString());
-            // Create the request to OpenWeatherMap, and open the connection
-            urlConnection = (HttpURLConnection) url.openConnection();
-            urlConnection.setRequestMethod("GET");
-            urlConnection.connect();
-
-            // Read the input stream into a String
-            InputStream inputStream = urlConnection.getInputStream();
-            StringBuffer buffer = new StringBuffer();
-            if (inputStream == null) {
-                // Nothing to do.
-                forecastJsonStr = null;
-            }
-            reader = new BufferedReader(new InputStreamReader(inputStream));
-
-            String line;
-            while ((line = reader.readLine()) != null) {
-                // Since it's JSON, adding a newline isn't necessary (it won't affect parsing)
-                // But it does make debugging a *lot* easier if you print out the completed
-                // buffer for debugging.
-                buffer.append(line + "\n");
-            }
-
-            if (buffer.length() == 0) {
-                // Stream was empty.  No point in parsing.
-                forecastJsonStr = null;
-            }
-
-            forecastJsonStr = buffer.toString();
-            Log.v(LOG_TAG, "Forecast JSON String: " + forecastJsonStr);
-            //Here was an offline error!
-            if (forecastJsonStr == null || forecastJsonStr.isEmpty()) {
-                Log.v(LOG_TAG, "Cann't receive the data. Check for your connection or may be you have incorrect query parameter");
-                return null;
-            }
-            Utils.getWeatherDataFromJson(mContext, forecastJsonStr, locationQuery);
-        } catch (IOException e) {
-            Log.e(LOG_TAG, "Error ", e);
-        } catch (JSONException e) {
-            Log.e(LOG_TAG, e.getMessage(), e);
-            e.printStackTrace();
-        } finally{
-            if (urlConnection != null) {
-                urlConnection.disconnect();
-            }
-            if (reader != null) {
-                try {
-                    reader.close();
-                } catch (final IOException e) {
-                    Log.e(LOG_TAG, "Error closing stream", e);
-                }
-            }
-        }
-
-        return null;
-    }
-}
