+++ Sunshine/app/src/main/res/layout-sw600dp/activity_main.xml
@@ -6,7 +6,7 @@
               android:baselineAligned="false"
               android:divider="?android:attr/dividerHorizontal"
               android:orientation="horizontal"
-              tools:context="com.example.android.sunshine.app.MainActivity">
+              tools:context="com.cdesign.sunshine.ui.MainActivity">
 
     <!--
     This layout is a two-pane layout for the Items master/detail flow.
@@ -14,7 +14,7 @@
 
     <fragment
         android:id="@+id/fragment_forecast"
-        android:name="com.example.android.sunshine.app.ForecastFragment"
+        android:name="com.cdesign.sunshine.ui.ForecastFragment"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="2"
