+++ Sunshine/app/src/main/res/layout/activity_detail.xml
@@ -4,5 +4,5 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/weather_detail_container"
     style="@style/match"
-    tools:context=".DetailActivity"
+    tools:context="com.cdesign.sunshine.ui.DetailActivity"
     tools:ignore="MergeRootFrame" />
