+++ Sunshine/app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment_forecast"
-    android:name="com.example.android.sunshine.app.ForecastFragment"
+    android:name="com.cdesign.sunshine.ui.ForecastFragment"
     style="@style/match"
-    tools:context="com.example.android.sunshine.app.ForecastFragment"
+    tools:context="com.cdesign.sunshine.ui.ForecastFragment"
     tools:layout="@android:layout/list_content" />
