+++ Sunshine/app/src/main/res/layout/fragment_main.xml
@@ -1,7 +1,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     style="@style/match"
-    tools:context=".ForecastFragment">
+    tools:context="com.cdesign.sunshine.ui.ForecastFragment">
 
     <ListView
         android:id="@+id/listview_forecast"
