+++ Sunshine/app/src/main/res/layout/list_item_forecast_today.xml
@@ -44,13 +44,11 @@
         <ImageView
             android:id="@+id/list_item_icon"
             style="@style/wrap"
-            android:layout_gravity="center_horizontal"
-            android:src="@drawable/ic_launcher"/>
+            android:layout_gravity="center_horizontal"/>
         <TextView
             android:id="@+id/list_item_forecast_txt"
             style="@style/wrap"
             android:layout_gravity="center_horizontal"
-            android:text="@string/clear"
             android:textColor="@color/white"
             android:fontFamily="sans-serif-condensed"
             android:textAppearance="?android:textAppearanceLarge"/>
