+++ Sunshine/app/src/main/res/menu/detail.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="com.example.android.sunshine.app.DetailActivity">
+      tools:context="com.cdesign.sunshine.ui.DetailActivity">
     <item
         android:id="@+id/action_settings"
         android:title="@string/action_settings"
