+++ Sunshine/app/src/main/res/menu/forecastfragment.xml
@@ -6,4 +6,9 @@
         android:title="@string/action_refresh"
         android:orderInCategory="100"
         app:showAsAction="never" />
+    <item
+        android:id="@+id/action_map"
+        android:title="@string/action_map"
+        android:orderInCategory="100"
+        app:showAsAction="never" />
 </menu>
