+++ Sunshine/app/src/main/res/menu/main.xml
@@ -1,15 +1,10 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity">
+    tools:context="com.cdesign.sunshine.ui.MainActivity">
     <item
         android:id="@+id/action_settings"
         android:title="@string/action_settings"
-        android:orderInCategory="100"
-        app:showAsAction="never" />
-    <item
-        android:id="@+id/action_map"
-        android:title="@string/action_map"
         android:orderInCategory="100"
         app:showAsAction="never" />
 </menu>
