+++ Sunshine/app/src/main/res/values/strings.xml
@@ -22,6 +22,18 @@
     <string name="pref_units_metric" translatable="false">metric</string>
     <string name="pref_units_imperial" translatable="false">imperial</string>
 
+    <string name="sync_account_type">sunshine.example.com</string>
+    <string name="content_authority">com.cdesign.sunshine</string>
+
+    <!-- Notifications -->
+    <string name="pref_last_notification">last_notification</string>
+    <string name="format_notification">Forecast: <xliff:g id="condition">%1$s</xliff:g> High: <xliff:g id="high">%2$s</xliff:g> Low: <xliff:g id="low">%3$s</xliff:g></string>
+    <string name="pref_notify_key" translatable="false">enable_notifications</string>
+    <string name="pref_notify_label">Weather Notifications</string>
+    <string name="pref_notify_true">Enabled</string>
+    <string name="pref_notify_false">Not Enabled</string>
+    <string name="pref_notify_default" translatable="false">true</string>
+
     <string name="today">Today</string>
     <!-- For labelling tomorrow's forecast [CHAR LIMIT=15] -->
     <string name="tomorrow">Tomorrow</string>
@@ -29,7 +41,7 @@
     <string name="high">81</string>
     <string name="low">68</string>
     <!-- Date format [CHAR LIMIT=NONE] -->
-    <string name="format_full_friendly_date"><xliff:g id="month">%1$s</xliff:g>, <xliff:g id="day">%2$s</xliff:g></string>
+    <string name="format_full_friendly_date"><xliff:g id="date">%1$s</xliff:g>, <xliff:g id="day">%2$s</xliff:g></string>
     <string name="format_temperature"><xliff:g id="temp">%1.0f</xliff:g>\u00B0</string>
 
     <!-- Wind in mph [CHAR LIMIT=25] -->
